Extract routes from main controller
