+++ src/server.ts
@@ -1,6 +1,9 @@
 import 'dotenv/config';
 
-import { PORT } from "./constants/config.constants";
+import { PORT } from "./constants/config";
+import DB from "./config/mongodb"
 import app from "./app";
+
+new DB();
 
 app.listen(PORT, () => console.log(`Server started on port ${PORT}`));
