+++ src/utils/permittedParams.ts
@@ -1,0 +1,4 @@
+import { pick  } from 'lodash';
+export const permittedParams = ( requestBody: Record<string, unknown>, permittedInputParams: readonly string[]): Record<string, unknown> => {
+  return pick(requestBody, permittedInputParams);
+};
